feat: remove `jump` states
To support `<<jump>>` commands, we were previously creating a state node for the command. To use
this, you'd have to call next on the preceding state, then call next again on the command node. This
update allows the preceding state node to hold the `next` action, removing the need for the command
to have its own state node,