feat: allow options to be selected properly
We were previously just adding the `next` event to all options in a parallel state. As it turns out,
there's no way to send an event to a single region in a parallel state. Instead, we needed to be
able to target a region with the event name. Ad accomodate, I've added IDs to the `next` event for
options so a specific option can now be selected using the `next::NODE_ID` syntax.

BREAKING CHANGE: sending `next` to an option no longer does anything.